FileStore: call op_apply_start when about to apply transaction
Previously, we called op_apply_start when we queue the transaction.  This
requires a sync to wait for the entire op queue to flush before performing
the sync.  Now, we simply wait for the currently in progress applies to
complete.  The op queueing ensures that this is enough to get a stable
commit point.

Not adequately tested.

Signed-off-by: Samuel Just <sam.just@inktank.com>