osd/: add pool min_size parameter for min acting set size
Otherwise, a pg might go acgtive with a single osd in the
acting set.  If that osd subsequently dies, we potentially
loose client writes.  Note: it's still possible for the
acting set to exceed min_size but fail to obey the spirit
of the user's crush settings (e.g., min_size is 2, but both
osds happen to be no the sam node).

Signed-off-by: Samuel Just <sam.just@inktank.com>