PG::activate: handle peer contigious with auth_log, but not primary
The added case covers a situation where a replica is not contiguous with
the auth_log, but is contiguous with this replica.  Reshuffling the
active set to handle this would be tricky, so instead we just go ahead
and backfill it anyway.  This is probably preferrable in any case since
the replica in question would have to be significantly behind.

Fixes: #7696
Signed-off-by: Samuel Just <sam.just@inktank.com>