librados: refactor watch/notify; return notify error code
Get rid of a level of intermediate classes with confusing names and put
the notify and notify finish logic in a single place so that it is easier
to follow and understand.

Pass the return value from the notify completion message to the caller.

Signed-off-by: Sage Weil <sage@redhat.com>
(cherry picked from commit 69612e7519f9fa73d6ac1c6358bfeac33e998774)

Conflicts:
	src/librados/RadosClient.cc