FileStore: Added a fast transaction apply scheme
Based on config options(filestore_do_fast_sync,filestore_do_force_sync),
a fast backend transaction apply and sync scheme can be triggered
within FileStore.This is only applicable for any filesystem for which
FileStore is writing a write ahead journal. The idea is to do a lot more
frequent syncfs call with less data to sync and an efficient way
to determinine the committing sequence number without pausing/unpausing
threadpool. This scheme won't do a file write everytime to save the last
committed seq number as we can get the last committing seq number
from journal header in case of write-ahead-journal scheme.
This scheme is very much needed if the journal size is very small, for ex
in case of configuring small NVRAM as journal and SSD as data store.

Signed-off-by: Somnath Roy <somnath.roy@sandisk.com>

Conflicts:
	src/os/filestore/FileStore.cc